DCaS Final Lab Portion
Code is now written for the Arduino Diecimila and the 6811 processor such that
the 6811 will output an oscillating signal at a speed dependent on a value
stored at memory address 0000 and the Arduino will light a single LED for each
of the three possible frequencies or no LEDs if the program is not running. The
program is commented and a video has been taken of it running, but a lab report
must still be written and circuit diagrams for the 4-10 decoder, NOT gate, and
D Flip-Flop that make up the hardware driver must be created.